Resolve IntelliJ IDEA inspection warnings and modernize code
- Replace deprecated HandlerMethodValidationException.getAllValidationResults() with getValueResults()
- Modernize lambdas to concise expression syntax in GlobalExceptionHandler
- Remove redundant @DecimalMin parameter and unused getter aliases in Book record
- Use Java 21 SequencedCollection getFirst() and remove unused import in BookServiceTest
- Replace Arrays.asList with List.of in BookControllerTest
- Type ParameterizedTypeReference in BookstoreApplicationIntegrationTest to eliminate raw Map warnings
- Remove redundant cast on allBooks.size() in BookService

+++ exercises/java/bookstore-api/src/main/java/com/example/bookstore/service/BookService.java
@@ -77,7 +77,7 @@ public class BookService {
             return List.of();
         }
         int start = (int) startLong;
-        int end = (int) Math.min(startLong + size, (long) allBooks.size());
+        int end = (int) Math.min(startLong + size, allBooks.size());
 
         return List.copyOf(allBooks.subList(start, end));
     }
